+++ src/Map.ts
@@ -737,12 +737,10 @@ export function getWitherable<K>(O: Ord<K>): Witherable2C<URI, K> & TraversableW
       for (let i = 0; i < len; i++) {
         const key = ks[i]
         const a = ta.get(key)!
-        fm = F.ap(
-          pipe(
-            fm,
-            F.map((m) => (b: B) => new Map(m).set(key, b))
-          ),
-          f(key, a)
+        fm = pipe(
+          fm,
+          F.map((m) => (b: B) => new Map(m).set(key, b)),
+          F.ap(f(key, a))
         )
       }
       return fm