+++ src/Tuple.ts
@@ -53,7 +53,7 @@ export function getApply<S>(S: Semigroup<S>): Apply2C<URI, S> {
     URI,
     _E: undefined as any,
     map,
-    ap: (fab, fa) => [fst(fab)(fst(fa)), S.concat(snd(fab), snd(fa))]
+    ap: (fa) => (fab) => [fst(fab)(fst(fa)), S.concat(snd(fab), snd(fa))]
   }
 }