+++ test/Tuple.ts
@@ -81,9 +81,9 @@ describe('Tuple', () => {
   })
 
   it('getApply', () => {
-    const apply = _.getApply(monoidString)
+    const A = _.getApply(monoidString)
     const double = (n: number): number => n * 2
-    assert.deepStrictEqual(apply.ap([double, 'a'], [1, 'b']), [2, 'ab'])
+    assert.deepStrictEqual(pipe([double, 'a'] as const, A.ap([1, 'b'] as const)), [2, 'ab'])
   })
 
   it('getApplicative', () => {